feat: migrate the huggingface det_callback [MLG-487]
