+++ examples/deepspeed/deepspeed_autotuning/hf_trainer_api/image_classification/image_classification.py
@@ -27,11 +27,9 @@ import numpy as np
 import torch
 import transformers
 from datasets import load_dataset
-from determined.hf_callback import (
+from determined.integrations.huggingface import (
     DetCallback,
     create_consistent_hf_args_for_deepspeed,
-    get_ds_config_path_from_args,
-    replace_ds_config_file_using_overwrites,
 )
 from determined.pytorch import dsat
 from PIL import Image
@@ -222,9 +220,9 @@ def parse_input_arguments(
     else:
         args = sys.argv[1:]
         args.extend(dict2args(training_arguments))
-        ds_config_path = get_ds_config_path_from_args(args)
+        ds_config_path = dsat.get_ds_config_path_from_args(args)
         if ds_config_path is not None:
-            replace_ds_config_file_using_overwrites(args, hparams)
+            dsat.replace_ds_config_file_using_overwrites(args, hparams)
             args = create_consistent_hf_args_for_deepspeed(args, ds_config_path)
         model_args, data_args, training_args = parser.parse_args_into_dataclasses(args)
