+++ examples/deepspeed/deepspeed_autotuning/hf_trainer_api/image_classification/startup-hook.sh
@@ -1,5 +1,3 @@
 #!/bin/bash
 
 pip install -r requirements.txt
-
-echo "running startup-hook.sh!"
