+++ examples/integrations/hf_trainer_api/image_classification.py
@@ -26,7 +26,6 @@ import numpy as np
 import torch
 import transformers
 from datasets import load_dataset
-from det_callback import DetCallback
 from PIL import Image
 from torch.utils.tensorboard import SummaryWriter
 from torchvision.transforms import (
@@ -52,6 +51,7 @@ from transformers.trainer_utils import get_last_checkpoint
 from transformers.utils.versions import require_version
 
 import determined as det
+from determined.integrations.huggingface import DetCallback
 
 """ Fine-tuning a 🤗 Transformers model for image classification"""
