+++ harness/determined/__init__.py
@@ -22,7 +22,6 @@ from determined._execution import (
 )
 from determined import errors
 from determined import util
-from determined import hf_callback
 
 # LOG_FORMAT is the standard format for use with the logging module, which is required for the
 # WebUI's log viewer to filter logs by log level.
