+++ harness/determined/integrations/__init__.py
@@ -1,0 +1,1 @@
+from determined.integrations.huggingface.hf_callback import DetCallback
