+++ harness/determined/integrations/huggingface/__init__.py
@@ -1,0 +1,4 @@
+from determined.integrations.huggingface.hf_callback import (
+    DetCallback,
+    create_consistent_hf_args_for_deepspeed,
+)
