+++ harness/determined/pytorch/dsat/__init__.py
@@ -2,6 +2,8 @@ from determined.pytorch.dsat._utils import (
     dsat_reporting_context,
     get_ds_config_from_hparams,
     overwrite_deepspeed_config,
+    get_ds_config_path_from_args,
+    replace_ds_config_file_using_overwrites,
 )
 from determined.pytorch.dsat._dsat_search_method import (
     BaseDSATSearchMethod,
